Renamed the folder for the Property Value Converter
