+++ src/Our.Umbraco.PropertyList/Bootstrap.cs
@@ -1,5 +1,5 @@
 ﻿using Newtonsoft.Json;
-using Our.Umbraco.PropertyList.Converters;
+using Our.Umbraco.PropertyList.ValueConverters;
 using Umbraco.Core;
 using Umbraco.Core.Sync;
 using Umbraco.Web.Cache;
