+++ src/Our.Umbraco.PropertyList/Our.Umbraco.PropertyList.csproj
@@ -82,7 +82,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Bootstrap.cs" />
-    <Compile Include="Converters\PropertyListValueConverter.cs" />
+    <Compile Include="ValueConverters\PropertyListValueConverter.cs" />
     <Compile Include="Models\PropertyListValue.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\VersionInfo.cs" />
